chore: split getEntityForEvalContext
